+++ packages/bueno/src/index.ts
@@ -3,6 +3,5 @@ export * from './values/value';
 export * from './values/number-value';
 export * from './values/boolean-value';
 export * from './values/string-value';
-export * from './values/record-value';
-export * from './values/array-value';
+export * from './values/complex-value';
 export * from './values/enum-value';