+++ packages/bueno/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.47](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.46...v0.1.0-alpha.47) (2020-11-11)
+
+
+### Features
+
+* **bueno:** add constrainTo option to StringValue ([8f56ff1](https://bitbucket.org/coveord/ui-kit/commits/8f56ff1f3330ac258840bff5d9b3fc24710ebf1c))
+
+
+
+
+
+# [0.1.0-alpha.45](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.44...v0.1.0-alpha.45) (2020-11-11)
+
+
+### Features
+
+* **bueno:** added ValueConfig to RecordValue ([9829eca](https://bitbucket.org/coveord/ui-kit/commits/9829ecae1d2f7be5ff06d62273025744b6470687))
+
+
+
+
+
+# [0.1.0-alpha.39](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.38...v0.1.0-alpha.39) (2020-11-09)
+
+
+### Features
+
+* **bueno:** add context message to #schema.validate, validate ResultsPerPage initialState ([d02dcee](https://bitbucket.org/coveord/ui-kit/commits/d02dceef7f5db7c89396bfdff0d3bb28120d2e72))
+
+
+
+
+
+# [0.1.0-alpha.38](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.37...v0.1.0-alpha.38) (2020-11-06)
+
+
+### Features
+
+* **bueno:** allow  RecordValue<string, ArrayValue> ([de086b0](https://bitbucket.org/coveord/ui-kit/commits/de086b07f4f984885510843c969dedec8921f7e9))
+
+
+
+
+
 # [0.1.0-alpha.35](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.34...v0.1.0-alpha.35) (2020-11-04)
 
 **Note:** Version bump only for package @coveo/bueno